requirements.txt
streamlit
langchain
upstash_vector
langchain_community
upstash_vector
openai
python-dotenv